client: refactor FormatStatusResponse to take bit mask for additional details
Adding a bool var for each additional details to print in the status
makes requires changing the function and all its callers and additionaly
makes certain invocations of the function hard to read without looking
at its signature and godoc comment due to the list of bools.

Refactor FormatStatusResponse to take a typed bit mask with the
additional details to print. This e.g. makes the invocation of a caller
not interested in any details more readable as:

    client.FormatStatusResponse(w, sr, client.StatusNoDetails)

compared to the previous version:

    client.FormatStatusResponse(w, sr, false, false, false, false, false)

Adding a new option for additional details now only requires adding an
additional mask const and its handling in FormatStatusResponse. Only
callers interested in this particular detail need to know about this
flag, existing callers don't need to be changed.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>