api, cli, daemon: add BPF map sizes to cilium status
When the dynamic BPF map sizing feature is used, for debugging, it would
be useful to know the actual map sizes without asking users to run
bpftool map show pinned ... or grepping logs of the agent.

Thus, report the BPF map sizes in and the dynamic size ratio in
`cilium status --verbose`.

Fixes #12643

Suggested-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>